#2 feat: Add ConfigSingleton method to retrieve minimum duration
Added `getMinDurationSeconds()` method to the `ConfigSingleton` class in `ConfigSingleton.cpp` to retrieve the minimum duration in seconds for processing an MP3 file. The minimum duration is read from the configuration file and is used in the `fileProcessor.cpp` module to validate the duration of the file before processing. If the duration is less than the minimum duration, the file is removed.