feat: Add fasterWhisper implementation and update main.cpp
The commit introduces a fasterWhisper implementation to improve the
execution speed of the sdrTrunkTranscriber application. The
include/whisperCPP.h file is removed, and a new include/fasterWhisper.h
header is added. The main.cpp file is updated to include the
fasterWhisper module instead of the whisperCPP module.

The addition of fasterWhisper enhances the application's transcription
processing capabilities, improving overall performance and efficiency.

Implements  enable local transcription with Whisper #13